python/requirements.txt
openai
pandas
numpy
pyarrow
python-dotenv
